Add PropertySet as weather property container
A PropertySet can be used by a weather controller to to store and manage (different) sets of values for same properties. It's used in the tests now since it is recommended for a weather controller to accept at least an initializing PropertySet in constructor. Such a PropertySet might be created with help of some builders. The do the parsing and construction of them.