Move API for several sky related constants into the SkyAppState
Some of these constants are currently hosted by the SkyGradient which isn't optimal but at least a common interface is used by all other AppStates - the SkyAppState.

These constants are strongly related to the sky mechanics and how the sky dome etc are placed in the world which should be specified by the SkyAppState as a central anchor somehow.