Move sky dome, stars and sun into Bucket.Sky
This makes that all those mentioned geometries will be rendered behind all other geometries, but we need a certain order on them in their infinity: sky dome, stars, sun. We assure this with a custom FixedOrderComparator using a user-defined data field in all important geometries that is used for sorting them correctly.