Prevent NPE in VertexBuffer.toString()
At the end of NativeObjectManager.deleteUnused() the cleaned up object is logged at FINEST. In case of the (cloned) VertexBuffer which has for some reason all null fields that toString() will throw a NPE during evaluation (in JLFBridge) of the log statement's parameter. By filtering the input to JLFBridge wth FINER (instead of ALL which would be the intended use) this can be circumvented.

--HG--
branch : feature-terrain