Add TerrainLUTGenerator as first preprocessor in sed.pre
Classes in sed.pre should be used to generate all assets that are needed by SED and statically producible like e.g. terrain lookup textures or maybe even texture atlases.

The TerrainLUTGenerator generates the a lookup table with is indexable by slope and altitude of a terrain point and delivers an ID that indicates a specific TerrainType there. From this type a shader could derive the texture or color to use.

--HG--
branch : feature-terrain