Add domain specific setType method to TerrainLUTBuilder
The old setType method is now called setTypeRect since it overrides a rectangle on pixel base. The new setType takes start and end slope/altitude as parameters which is far more usable for the TerrainLUTGenerator since he has not to bother with conversion of values.

All fields of TextureMapBuilder are now private, getters supplied.

--HG--
branch : feature-terrain