Initial draft for a (layered) settings system
The Settings interface is basically designed for use in all components that have nothing to do with settings management (as a settings dialog would have). It provides simple getter and setter with some integrated type conversion/parsing.

The LayeredSettingsManager implements this interface and provides a layer mechanism behind the scenes. For each setting queried it searches all layers starting with the most recently attached one and returns early on match. Changed settings (via setX()) will be persisted in the first WritableLayer found.