Vary rain drop length by precipitation.intensity
A poor implementation of the precipitation's intensity is done by varying the drop length since the drop length has some visual impact and influences the "visual density".
It should actually should manifest itself mostly in the drop density per volume but that's difficult since the number of particles is not modifiable after initialization yet.